requirements.txt
# HTTP and HTML parsing
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
demjson3>=3.0.6

# Environment
python-dotenv>=1.0.0

# Audio tagging (extract.py)
mutagen>=1.47.0

# archive.org upload (Step 7)
internetarchive>=4.0.0
